\name{AbbrevToNames}
\alias{AbbrevToNames}

\title{
Primarily for internal use, but this function takes the three letter abbreviation for each franchise and returns the city and nickname.
}
\description{
Primarily for internal use, but this function takes the three letter abbreviation for each franchise and returns the city and nickname.
}
\usage{
AbbrevToNames(Abbrev)
}
\arguments{
  \item{Abbrev}{Character string of the abbreviation.}
}
\details{
}
\value{\item{string}{Returns the NBA franchise in a character string.}}
\\examples{
AbbrevToNames("BOS")

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{Names}% use one of  RShowDoc("KEYWORDS")
\keyword{Abbreviations}% __ONLY ONE__ keyword per line
\keyword{Franchise}
\keyword{City}
