\name{about_ed}
\alias{about_ed}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Ed Bartels mean true shooting percentage
}
\description{
This function returns Ed Bartels' mean true shooting percentage in a given year. If the year is not one in which Ed Bartels played, the function will return "NaN."
}
\usage{
about_ed(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Year from the nba_data dataset
}
}
\author{
Amanda
}
\examples{
x=1955
## The function is currently defined as
function (x) 
{1995
  }
}
