% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Age_max.r
\name{Age_max}
\alias{Age_max}
\title{A Age max function}
\usage{
Age_max(df, yr)
}
\arguments{
\item{df}{A dataframe}

\item{yr}{year}
}
\description{
This function allows you calculate the player with the largest age in a given year
}
