% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allstars.R
\docType{data}
\name{allstars}
\alias{allstars}
\title{List of NBA All Star players by 2018}
\format{A data frame with 1,663 rows and 2 variables:
\describe{
  \item{name}{name}
  \item{year}{selected year}
}}
\source{
\url{https://en.wikipedia.org/wiki/List_of_NBA_All-Stars}
}
\usage{
data(allstars)
}
\description{
List of NBA All Star players by 2018
}
\keyword{datasets}
