% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{cor_plot}
\alias{cor_plot}
\title{A Correlation Function}
\usage{
cor_plot(df, year)
}
\arguments{
\item{df}{the data frame}

\item{year}{the specific year}
}
\description{
Creates a correlation matrix of df
}
\examples{
cor_plot()
}
\keyword{corr_plot}
