% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_data.R
\name{corr_data}
\alias{corr_data}
\title{This function selects all numeric variables for a given year and creates a correlogram.}
\usage{
corr_data(x, y)
}
\arguments{
\item{x}{deienfes the dataset}

\item{y}{difnes the year}
}
\description{
This function selects all numeric variables for a given year and creates a correlogram.
}
\examples{
corr_data()
}
\keyword{correlogram}
