% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\title{corr matrix function}
\usage{
corr_matrix(year = 1990)
}
\arguments{
\item{year}{What year are you interested in? Defaults to 1990.}
}
\description{
This function creates a correlation matrix of all continuous numeric variables in a given year for the nba dataset
}
\examples{
corr_matrix()
}
\keyword{correlation}
\keyword{matrix,}
