% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eldest_player.R
\name{eldest_player}
\alias{eldest_player}
\title{eldest player function}
\usage{
eldest_player(year = 1990)
}
\arguments{
\item{year}{What year are you interested in? Defaults to 1990.}
}
\description{
This function allows you find the oldest NBA player in a given year.
}
\examples{
eldest_player()
}
\keyword{eldest}
