% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_percent.R
\name{fg_percent}
\alias{fg_percent}
\title{fg_percent function}
\usage{
fg_percent(year = 1990)
}
\arguments{
\item{year}{What year are you interested in? Defaults to 1990.}
}
\description{
This function allows you to check Michael Jordan's field goal percentage in a given year.
}
\examples{
fg_percent()
}
\keyword{FG}
