\name{GetSeason}
\alias{GetSeason}
\title{
Scrapes the league stats for the season
}
\description{
This function returns the stats of individual teams for any single year within the past 10 years.
}
\usage{
GetSeason(year)
}
\arguments{
  \item{year}{a character value that defines the year of the league statistic table.}
}

\examples{
get.season("2015")
}

\keyword{Season}% use one of  RShowDoc("KEYWORDS")
\keyword{Stats}% __ONLY ONE__ keyword per line
