% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_points.R
\name{highest_points}
\alias{highest_points}
\title{A Highest Points Function}
\usage{
highest_points(year)
}
\arguments{
\item{year}{Year of interest}
}
\description{
This function allows you find out the player with the highest points of the input year.
}
\examples{
highest_points()
}
\keyword{points}
