% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_free_throw_attempted.R
\name{most_free_throw_attempted}
\alias{most_free_throw_attempted}
\title{Free throws attempted}
\usage{
most_free_throw_attempted(df, year)
}
\arguments{
\item{df}{the data frame}

\item{year}{the specific year}
}
\description{
Finds most free throws in a year
}
\examples{
most_free_throw_attempted()
}
\keyword{free_throw}
