% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_points.R
\name{most_points}
\alias{most_points}
\title{This function returns the name and points of the highest scoring player for a given year.}
\usage{
most_points(x, y)
}
\arguments{
\item{x}{deienfes the dataset}

\item{y}{difnes the year}
}
\description{
This function returns the name and points of the highest scoring player for a given year.
}
\examples{
corr_data()
}
