\name{NamesToAbbrev}
\alias{NamesToAbbrev}
\title{
Primarily for internal use, but this function takes the city and nickname for each franchise and returns the three letter abbreviation.
}
\description{
Primarily for internal use, but this function takes the city and nickname for each franchise and returns the three letter abbreviation.
}
\usage{
NamesToAbbrev(Name)
}
\arguments{
  \item{Name}{Character string of the franchise.}
}
\details{
}
\value{\item{string}{Returns the three letter abbreviation in a character string.}}
\examples{
NamesToAbbrev("Boston Celtics")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{Names}% use one of  RShowDoc("KEYWORDS")
\keyword{Abbreviations}% __ONLY ONE__ keyword per line
\keyword{Franchise}
\keyword{City}
