% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_data_description.R
\docType{data}
\name{nba}
\alias{nba}
\title{NBA Season Stats for Players}
\format{
A data frame with 24691 rows and 52 variables:
\describe{
\item{Year}{Year representing the NBA season}
\item{Player}{Name of player}
...
}
}
\usage{
nba
}
\description{
A dataset containing the stats of NBA players across several years
}
\keyword{datasets}
