% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA_cor.R
\name{NBA_cor}
\alias{NBA_cor}
\title{NBA correlation matrix for numeric continuous variables of a year}
\usage{
NBA_cor(x)
}
\arguments{
\item{Year}{}
}
\description{
Take year input, returns correlation matrix for numeric continuous variables.
}
\examples{
NBA_cor(1998)
}
\keyword{correlation}
\keyword{matrix}
