% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA_most_point.R
\name{NBA_most_point}
\alias{NBA_most_point}
\title{NBA Most Point Player of a year}
\usage{
NBA_most_point(x)
}
\arguments{
\item{Year}{}
}
\description{
Take year input, returns players who scored the most points.
}
\examples{
NBA_most_point(1998)
}
\keyword{most_point}
