% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA_oldest.R
\name{NBA_oldest}
\alias{NBA_oldest}
\title{NBA Oldest Player of a year}
\usage{
NBA_oldest(x)
}
\arguments{
\item{Year}{}
}
\description{
Take year input, returns oldest players.
}
\examples{
NBA_oldest(1998)
}
\keyword{oldest}
