% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{This function returns the name and age of the oldest player for a given year.}
\usage{
oldest_player(x, y)
}
\arguments{
\item{x}{deienfes the dataset}

\item{y}{difnes the year}
}
\description{
This function returns the name and age of the oldest player for a given year.
}
\examples{
corr_data()
}
