% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_stats.R
\name{oldest_stats}
\alias{oldest_stats}
\title{Oldest}
\usage{
oldest_stats(df, year)
}
\arguments{
\item{df}{the data frame}

\item{year}{the specific year}
}
\description{
Oldest player in a year
}
\examples{
oldest_stats()
}
\keyword{oldest}
