% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_data.R
\docType{data}
\name{player_data}
\alias{player_data}
\title{Datasets of NBA players since 1950}
\format{A data frame with 4,550 rows and 8 variables:
\describe{
  \item{name}{name}
  \item{year_start}{start year of the player's carrer}
  \item{end_year}{end year of the player's carrer}
  \item{position}{position of the player}
  \item{height}{height of the player, in ft-in}
  \item{weight}{weight of the player, in lb}
  \item{birth_date}{birth day of the player}
  \item{college}{college of the player}
}}
\source{
\url{https://www.kaggle.com/drgilermo/nba-players-stats}
}
\usage{
data(player_data)
}
\description{
A dataset of NBA player's profile.
}
\keyword{datasets}
