% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{player_piclink}
\alias{player_piclink}
\title{Get the link of given player's picture}
\usage{
player_piclink(player)
}
\arguments{
\item{player}{player name (character string)}
}
\value{
a character string which is the link of player's picture
}
\description{
Function to built a link to given player's picture from Basketball Reference
}
\examples{
player_piclink("Stephen Curry")
}
\references{
\url{https://www.basketball-reference.com}
}
\seealso{
\code{\link{season_players}}
}
\author{
Ozan Adiguzel
}
\keyword{nba}
\keyword{pergame}
\keyword{stats}
