% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{player_totals}
\alias{player_totals}
\title{Scrape regular season stats totals for given player}
\usage{
player_totals(player, start_year)
}
\arguments{
\item{player}{player name (character string)}

\item{start_year}{start year of the season (numeric value from 1949 to 2017)}
}
\value{
a data frame of season stats totals
}
\description{
Function to scrape regular season stats totals for given player from Basketball Reference
}
\note{
Glossary: \url{https://www.basketball-reference.com/about/glossary.html}
}
\examples{
player_totals("Stephen Curry", 2017)
}
\references{
\url{https://www.basketball-reference.com}
}
\seealso{
\code{\link{season_players}}
}
\author{
Ozan Adiguzel
}
\keyword{nba}
\keyword{stats}
\keyword{totals}
