% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\docType{data}
\name{players}
\alias{players}
\title{Datasets of NBA players since 1950}
\format{A data frame with 3,922 rows and 7 variables:
\describe{
  \item{Player}{name}
  \item{height}{height of the player, in cm}
  \item{weight}{weight of the player, in kg}
  \item{college}{college of the player}
  \item{born}{born yearof the player}
  \item{birth_city}{birth city of the player}
  \item{birth_state}{birth state of the player}
}}
\source{
\url{https://www.kaggle.com/drgilermo/nba-players-stats}
}
\usage{
data(players)
}
\description{
A dataset about NBA player's profile.
}
\keyword{datasets}
