% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTS_max.R
\name{PTS_max}
\alias{PTS_max}
\title{A most Points function}
\usage{
PTS_max(df, yr)
}
\arguments{
\item{df}{A dataframe}

\item{yr}{year}

\item{points}{max points for specific year}
}
\description{
This function allows you calculate the player with the most points in a given year
}
