% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salaries.R
\docType{data}
\name{salaries}
\alias{salaries}
\title{Datasets of NBA players' salaries}
\format{A data frame with 8,691 rows and 4 variables:
\describe{
  \item{season}{season year}
  \item{name}{players' name}
  \item{salary}{salary, in US dollar}
  \item{salary_adj}{salary adjusted for inflation, in US dollar}
}}
\source{
\url{https://hoopshype.com/salaries/players/2017-2018/}
}
\usage{
data(salaries)
}
\description{
A dataset of NBA players' salaries since 2000-2001 season to 2017-2018 season.
}
\keyword{datasets}
