% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{season_players}
\alias{season_players}
\title{Get list of players by season}
\usage{
season_players(start_year)
}
\arguments{
\item{start_year}{start year of the season (numeric value from 1949 to 2017)}
}
\value{
a list of players
}
\description{
Function to get list of players
}
\examples{
season_players(2017)
}
\references{
\url{https://www.basketball-reference.com}
}
\author{
Ozan Adiguzel
}
\keyword{nba}
\keyword{player}
\keyword{stats}
