% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{season_ranking}
\alias{season_ranking}
\title{Scrape regular season rankings}
\usage{
season_ranking(start_year)
}
\arguments{
\item{start_year}{start year of the season (numeric value from 1976 to 2017)}
}
\value{
a data frame of regular season ranking
}
\description{
Function to scrape regular season rankings from Wikipedia
}
\examples{
season_ranking(2017)

}
\references{
\url{https://www.wikipedia.org}
}
\author{
Ozan Adiguzel
}
\keyword{nba}
\keyword{ranking}
