% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{season_totals}
\alias{season_totals}
\title{Scrape regular season stats totals for all players}
\usage{
season_totals(start_year)
}
\arguments{
\item{start_year}{start year of the season (numeric value from 1949 to 2017)}
}
\value{
a data frame of season stats totals
}
\description{
Function to scrape regular season stats totals for all players from Basketball Reference
}
\note{
Glossary: \url{https://www.basketball-reference.com/about/glossary.html}
}
\examples{
season_totals(2017)

}
\references{
\url{https://www.basketball-reference.com}
}
\author{
Ozan Adiguzel
}
\keyword{nba}
\keyword{stats}
\keyword{totals}
