% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corr_matrix.R
\name{Corr_matrix}
\alias{Corr_matrix}
\title{A correlation function}
\usage{
Corr_matrix(df, yr)
}
\arguments{
\item{df}{A dataframe}

\item{yr}{year}

\item{corr}{corr for specific year}
}
\description{
This function allows you calculate the correlation of players in a given year
}
