% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\title{A Correlation Matrix Function}
\usage{
corr_matrix(data)
}
\arguments{
\item{data}{input data}
}
\description{
This function allows you find out correlation among the numeric variables.
}
\examples{
corr_matrix()
}
\keyword{correlation}
