% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{A Oldest Player Function}
\usage{
oldest_player(year)
}
\arguments{
\item{year}{Year of interest}
}
\description{
This function allows you find out the oldest player of the input year.
}
\examples{
oldest_player()
}
\keyword{player}
