% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons_stats.R
\docType{data}
\name{seasons_stats}
\alias{seasons_stats}
\title{Datasets of NBA players' stats since 1950}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24691 rows and 52 columns.}
\source{
\url{https://www.kaggle.com/drgilermo/nba-players-stats}
  \url{https://www.basketball-reference.com/about/glossary.html}
}
\usage{
data(seasons_stats)
}
\description{
A dataset of NBA players' stats.
}
\keyword{datasets}
