% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons_stats.R
\docType{data}
\name{seasons_stats}
\alias{seasons_stats}
\title{Datasets of NBA players' stats since 1950}
\format{A data frame with 24,691 rows and 52 variables:
\describe{
  \item{Year}{year}
  \item{Player}{name the player}
  \item{Pos}{position of the player}
  \item{position}{position of the player}
  \item{Age}{Age of the player}
  \item{Tm}{Team, in 3 letters}
  ... \n
  see below for details \n
  \url{https://www.basketball-reference.com/about/glossary.html}
}}
\source{
\url{https://www.kaggle.com/drgilermo/nba-players-stats}
}
\usage{
data(seasons_stats)
}
\description{
A dataset about NBA players' stats.
}
\keyword{datasets}
